requirements.txt
torch
transformers
peft
datasets
accelerate
trl
gradio
pyyaml
ruff
